list_cameras.py
import cv2

def list_cameras():
    available_cameras = []
    for i in range(5):
        cap = cv2.VideoCapture(i)
        if cap.isOpened():
            available_cameras.append(i)
            cap.release()
    return available_cameras

cameras = list_cameras()
print(f"Available cameras: {cameras}")