requirements.txt
opencv-python>=4.6.0
pyserial>=3.5
